+++ mek_data_class_generator/lib/src/writers/fields_class_writer.dart
@@ -1,10 +1,14 @@
 import 'package:analyzer/dart/element/element.dart';
+import 'package:json_annotation/json_annotation.dart';
 import 'package:mek_data_class_generator/src/configs.dart';
 import 'package:mek_data_class_generator/src/specs.dart';
 import 'package:mek_data_class_generator/src/utils.dart';
 import 'package:mek_data_class_generator/src/writers/writer.dart';
+import 'package:source_gen/source_gen.dart';
 
 class FieldsClassWriter extends Writer {
+  static final _jsonSerializableType = TypeChecker.fromRuntime(JsonSerializable);
+
   final Config config;
 
   FieldsClassWriter({
@@ -13,7 +17,11 @@ class FieldsClassWriter extends Writer {
     required List<FieldSpec> fieldSpecs,
   }) : super(classSpec: classSpec, fieldSpecs: fieldSpecs);
 
-  String? buildDataField(FieldSpec fieldSpec) {
+  String buildFieldPath(FieldSpec fieldSpec, bool hasFieldMap) {
+    return hasFieldMap ? '\$_path\${_get(\'${fieldSpec.name}\')}' : '\${_path}${fieldSpec.name}';
+  }
+
+  String? buildDataField(FieldSpec fieldSpec, String fieldPath) {
     final fieldClassElement = fieldSpec.element.type.element2;
     if (fieldClassElement is! ClassElement) return null;
 
@@ -24,26 +32,39 @@ class FieldsClassWriter extends Writer {
     if (!fieldClassSpec.createFieldsClass) return null;
 
     final keysClassName = '${fieldClassElement.name}Fields';
-    return '$keysClassName get ${fieldSpec.name} => $keysClassName(\'\${_path}${fieldSpec.name}.\');';
+    return '$keysClassName get ${fieldSpec.name} => $keysClassName(\'$fieldPath.\');';
   }
 
-  String buildField(FieldSpec fieldSpec) {
-    final result = buildDataField(fieldSpec);
+  String buildField(FieldSpec fieldSpec, bool hasFieldMap) {
+    final fieldPath = buildFieldPath(fieldSpec, hasFieldMap);
 
-    return result ?? 'String get ${fieldSpec.name} => \'\${_path}${fieldSpec.name}\';';
+    final result = buildDataField(fieldSpec, fieldPath);
+
+    return result ?? 'String get ${fieldSpec.name} => \'$fieldPath\';';
   }
 
   @override
   Iterable<String> writeClasses() sync* {
+    final jsonSerializable = _jsonSerializableType.firstAnnotationOf(classSpec.element);
+    final hasFieldMap = ConstantReader(jsonSerializable).peek('createFieldMap')?.boolValue ?? false;
+
     final className = '${classSpec.element.name}Fields';
+
     yield '''class $className {
     final String _path;
 
   const $className([this._path = '']);
-
-  ${fieldSpecs.map(buildField).join('\n')}    
+  
+  ${fieldSpecs.map((e) => buildField(e, hasFieldMap)).join('\n')}    
   
   String toString() => _path.isEmpty ? '$className()' : _path;
-}''';
+''';
+    if (hasFieldMap) {
+      yield '''
+
+  String _get(String key) => _\$${classSpec.self.name}FieldMap[key]!;
+''';
+    }
+    yield '\n}';
   }
 }